add uniform random element generation
As suggested in the lecture notes by Alexander Hulpke, do this with a
modified element test. Compared to sift!, the algorithm starts at the
bottom of the chain (starting from g == 1) and works its way upwards.

Example distribution for 1000 samples and a permutation group of order
10, with stabilizer chain depth == 2:

10-element Vector{Tuple{Tuple{Int64, Permutation}, Int64}}:
 ((1, (3,6,4,7,5)), 934)
 ((1, (1,2)(3,6,4,7,5)), 1054)
 ((1, (3,5,7,4,6)), 1005)
 ((1, ()), 1066)
 ((1, (3,7,6,5,4)), 1017)
 ((1, (1,2)(3,7,6,5,4)), 970)
 ((1, (1,2)(3,5,7,4,6)), 998)
 ((1, (3,4,5,6,7)), 1008)
 ((1, (1,2)), 955)
 ((1, (1,2)(3,4,5,6,7)), 993)

The tests take 50000 samples for each permutation group up to order 30,
and verify uniform distribution (up to a confidence of 92.5%).